+++ Automate/Framework/Machines/Objects/BeeHouseMachine.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using Microsoft.Xna.Framework;
 using StardewValley;
+using StardewValley.TerrainFeatures;
 using SObject = StardewValley.Object;
 
 namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
@@ -66,13 +67,13 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
             int addedPrice = 0;
             if (this.Location is Farm)
             {
-                Crop flower = Utility.findCloseFlower(this.Tile);
+                Crop flower = this.FindCloseFlower(this.Location, this.Tile);
                 if (flower != null)
                 {
                     string[] flowerData = Game1.objectInformation[flower.indexOfHarvest].Split('/');
                     prefix = flowerData[0];
                     addedPrice = Convert.ToInt32(flowerData[1]) * 2;
-                    if(!this.HoneyTypes.TryGetValue(flower.indexOfHarvest, out type))
+                    if (!this.HoneyTypes.TryGetValue(flower.indexOfHarvest, out type))
                         type = SObject.HoneyType.Wild;
                 }
             }
@@ -111,6 +112,40 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
             machine.minutesUntilReady = 2400 - Game1.timeOfDay + 4320;
             machine.readyForHarvest = false;
             machine.showNextIndex = false;
+        }
+
+        /// <summary>Get the closest flower within range of the given beehive.</summary>
+        /// <param name="location">The bee hive's location.</param>
+        /// <param name="tile">The bee hive's tile coordinate.</param>
+        /// <remarks>This logic is duplicated from <see cref="Utility.findCloseFlower"/>, but allows for any location instead of being hardcoded to the farm.</remarks>
+        private Crop FindCloseFlower(GameLocation location, Vector2 tile)
+        {
+            // use default game logic if possible
+            if (location is Farm)
+                return Utility.findCloseFlower(tile);
+
+            // handle flowers in custom locations (e.g. Farm Expansion)
+            {
+                Queue<Vector2> queue = new Queue<Vector2>();
+                HashSet<Vector2> visited = new HashSet<Vector2>();
+                queue.Enqueue(tile);
+                for (int i = 0; i <= 150 && queue.Count > 0; i++)
+                {
+                    // check for fully-grown tile on the current tile
+                    Vector2 curTile = queue.Dequeue();
+                    if (location.terrainFeatures.ContainsKey(curTile) && location.terrainFeatures[curTile] is HoeDirt dirt && dirt.crop != null && dirt.crop.programColored && dirt.crop.currentPhase >= dirt.crop.phaseDays.Count - 1 && !dirt.crop.dead)
+                        return dirt.crop;
+
+                    // try surrounding tiles
+                    foreach (Vector2 nextTile in Utility.getAdjacentTileLocations(curTile))
+                    {
+                        if (!visited.Contains(nextTile))
+                            queue.Enqueue(nextTile);
+                    }
+                    visited.Add(curTile);
+                }
+                return null;
+            }
         }
     }
 }